refactor(panel): extract chooser-tile launch handoff into useChooserHandoff
Amp-Thread-ID: https://ampcode.com/threads/T-019e2394-e4cc-73b9-a95c-23a5a5143a41
